category service tests modified update and delete category unit tests added to project

+++ src/SuperMarket.Services/Categories/CategoryAppService.cs
@@ -46,12 +46,22 @@ namespace SuperMarket.Services.Categories
         {
             var category = _repository.FindById(id);
             var isCategoryWithProduct = _repository.IsCategoryWithProduct(id);
+            var isCategoryExist = _repository.IsCategoryExist(id);
+            
             if (isCategoryWithProduct)
             {
                 throw new CategoryContainProductException();
             }
-            _repository.Delete(category);
-            _unitOfWork.Commit();
+            if(isCategoryExist)
+            {
+                _repository.Delete(category);
+                _unitOfWork.Commit();
+            }
+            else
+            {
+                throw new ThereIsNoCtegoryWithThisIdException();
+            }
+         
         }
 
         public IList<GetCategoryDto> GetAll()
@@ -63,13 +73,22 @@ namespace SuperMarket.Services.Categories
         {
             Category category = _repository.FindById(id);
             bool isNameDuplicate = _repository.IsExistCategoryName(dto.Name);
+            var iscategoryexist = _repository.IsCategoryExist(id);
             if (isNameDuplicate)
             {
                 throw new DuplicateCategoryNameException();
             }
-            category.Name = dto.Name;
-            _repository.Update(category);
-            _unitOfWork.Commit();
+            if(iscategoryexist)
+            {
+                category.Name = dto.Name;
+                _repository.Update(category);
+                _unitOfWork.Commit();
+            }
+            else
+            {
+                throw new ThereIsNoCtegoryWithThisIdException();
+            }
+        
         }
     }
 }
